Move logging and commandline switch configuration to own files
