+++ Cql/PackagerCLI/Program.cs
@@ -7,21 +7,15 @@
  */
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-using System.Diagnostics;
 using System.Globalization;
-using Hl7.Cql.Abstractions.Exceptions;
 using Hl7.Cql.CodeGeneration.NET;
 using Hl7.Cql.Packager.Hosting;
-using Hl7.Cql.Packager.Logging;
 using Hl7.Cql.Packaging;
 using Hl7.Cql.Packaging.PostProcessors;
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Serilog;
-using Serilog.Events;
 
 namespace Hl7.Cql.Packager;
 
@@ -50,11 +44,11 @@ public class Program
 
     public static IHostBuilder CreateHostBuilder(string[] args) =>
         Host.CreateDefaultBuilder()
-            .ConfigureAppConfiguration((context, config) => config.ConfigurePackagerCliApp(args))
-            .ConfigureLogging((context, logging) => ConfigureLogging(logging, context.Configuration))
+            .ConfigureAppConfiguration((context, config) => config.AddPackagerCliCommandLineSwitches(args))
+            .ConfigureLogging((context, logging) => logging.AddPackagerCLiLogging(context.Configuration))
             .ConfigureServices((context, services) =>
                                    services
-                                       .ConfigurePackagerCliOptions(context.Configuration)
+                                       .AddPackagerCliOptions(context.Configuration)
                                        .AddPackagerCliServices());
 
     private static string Usage { get; }
@@ -83,66 +77,6 @@ public class Program
 
         static string Optional(string s) => $"[{s}]";
     }
-
-    public static void ConfigureLogging(
-        ILoggingBuilder logging,
-        IConfiguration configuration)
-    {
-        logging.ClearProviders();
-
-        bool enableDebugLogging = Debugger.IsAttached;
-        enableDebugLogging = enableDebugLogging || configuration.GetSwitchMappedValue<bool>(CqlToResourcePackagingOptions.ArgNameLogDebugEnabled);
-        var minLogLevel = enableDebugLogging ? LogLevel.Trace : LogLevel.Information;
-
-        bool shouldClearLog = !configuration.GetSwitchMappedValue<bool>(CqlToResourcePackagingOptions.ArgNameLogDontClear);
-
-        logging.AddFilter(level => level >= minLogLevel);
-
-        logging.AddCleanConsole(opt =>
-        {
-            // opt.NoColor = true;
-        });
-
-        var logFile = Path.Combine(".", "build.log");
-
-        if (shouldClearLog)
-            File.WriteAllText(logFile, ""); // Create or clear the log file
-                                            //File.Delete(logFile);
-        else
-            File.OpenText(logFile).Close(); // Touch the file
-
-        Log.Logger = new LoggerConfiguration()
-            .Enrich.FromLogContext()
-            .MinimumLevel.Is(MapToSeriLogLogEventLevel(minLogLevel)!.Value)
-            .WriteTo.File(
-                logFile,
-                outputTemplate: "____{NewLine}{Level:u4}: {Message:lj}{NewLine}{Exception}",
-                formatProvider: CultureInfo.InvariantCulture)
-            .CreateLogger();
-        logging.AddSerilog();
-    }
-
-    private static LogEventLevel? MapToSeriLogLogEventLevel(LogLevel logLevel) =>
-        logLevel switch
-        {
-            // @formatter: off
-            /* 0 */
-            LogLevel.Trace => /* 0 */ LogEventLevel.Verbose,
-            /* 1 */
-            LogLevel.Debug => /* 1 */ LogEventLevel.Debug,
-            /* 2 */
-            LogLevel.Information => /* 2 */ LogEventLevel.Information,
-            /* 3 */
-            LogLevel.Warning => /* 3 */ LogEventLevel.Warning,
-            /* 4 */
-            LogLevel.Error => /* 4 */ LogEventLevel.Error,
-            /* 5 */
-            LogLevel.Critical => /* 5 */ LogEventLevel.Fatal,
-            /* 6 */
-            LogLevel.None => /* n/a */ null,
-            // @formatter: on
-            _ => throw new UnsupportedSwitchCaseError(logLevel, typeof(LogLevel).FullName).ToException(),
-        };
 
     private static int Run(IHostBuilder hostBuilder)
     {
